do the filtering in python and use the non-Approx wrapper

+++ src/sentry/migrations/0572_rm_cficache_symcache.py
@@ -3,7 +3,7 @@
 from django.db import migrations
 
 from sentry.new_migrations.migrations import CheckedMigration
-from sentry.utils.query import RangeQuerySetWrapperWithProgressBarApprox
+from sentry.utils.query import RangeQuerySetWrapperWithProgressBar
 
 # There is a bunch of orphaned `File`s still being stored that will be cleaned up.
 # They have the following `type`s:
@@ -21,10 +21,9 @@ def rm_cficache_symcache(apps, schema_editor):
     # > it is better to iterate over the entire table instead of using a filter
     # Also, there is no index on `type` anyway.
     # However, we also do not want to stream down every object into Python land.
-    for file in RangeQuerySetWrapperWithProgressBarApprox(
-        File.objects.filter(type__in=TYPES_TO_DELETE)
-    ):
-        file.delete()
+    for file in RangeQuerySetWrapperWithProgressBar(File.objects.all()):
+        if file.type in TYPES_TO_DELETE:
+            file.delete()
 
 
 class Migration(CheckedMigration):
